+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Some Python cleaning about handling file path, related to CFG file for instance
 * When running Lizmap Web Client 3.6, the QGIS server must be OK before going in other tabs of the plugin
 * Disable or replace graphical components if Qt Webkit is not available on the computer.
+* Fix the launching of the plugin on QGIS 3.10
 
 ## 3.11.1 - 2023-03-21